requirements.txt
picamera2
Pillow
pyproj
numpy
pandas
